Refactor layers state into custom hook for testing
