+++ frontend/src/app/editor/page.tsx
@@ -2,7 +2,7 @@
 
 import Konva from "konva";
 import dynamic from "next/dynamic";
-import React, { useState, useEffect, useCallback, useRef } from "react";
+import React, { useState, useEffect, useRef } from "react";
 
 import { MOVE_TOOL } from "@/components/editor/tools/move";
 import { ToolConfiguration } from "@/models/editor/tools/toolConfiguration";
@@ -13,12 +13,6 @@ import { LayersManagement } from "@/components/editor/layers/layersManagement";
 import { Menu } from "@/components/menu/menu";
 import { Toolbar } from "@/components/editor/toolbar/toolbar";
 
-import {
-  Layer,
-  LayerId,
-  LayerUpdateCallback,
-} from "@/models/editor/layers/layer";
-import { useUndoRedo } from "@/hooks/useUndoRedo";
 import { TOOLS, TOOLS_INITIAL_STATE } from "@/models/editor/utils/tools";
 
 import {
@@ -26,6 +20,7 @@ import {
   CanvasState,
   EventHandlers,
 } from "@/components/editor/editorContext";
+import useLayersState from "@/hooks/useLayersState";
 
 const Canvas = dynamic(() => import("@/components/editor/canvas"), {
   ssr: false,
@@ -46,17 +41,6 @@ for (const tool of Object.values(TOOLS)) {
 }
 
 export default function EditorPage() {
-  const {
-    state: layers,
-    setState: setLayers,
-    setVirtualState: setVirtualLayers,
-    commitVirtualState: commitVirtualLayers,
-    undo,
-    redo,
-    canUndo,
-    canRedo,
-  } = useUndoRedo(Array<Layer>());
-
   const [nameSelectedTool, setNameSelectedTool] = useState<string>(
     MOVE_TOOL.name,
   );
@@ -71,6 +55,20 @@ export default function EditorPage() {
 
   const canvasContainerRef = useRef<HTMLDivElement>(null);
   const [canvasSize, setCanvasSize] = useState({ width: 0, height: 0 });
+
+  const {
+    layers,
+    setLayers,
+    setVirtualLayers,
+    commitVirtualLayers,
+    undo,
+    redo,
+    canUndo,
+    canRedo,
+
+    updateLayer,
+    editSelectedLayers,
+  } = useLayersState();
 
   useEffect(() => {
     const container = canvasContainerRef.current;
@@ -119,55 +117,6 @@ export default function EditorPage() {
       x: (stagePosition.x - canvas.position.x) / canvas.scale,
       y: (stagePosition.y - canvas.position.y) / canvas.scale,
     };
-  };
-
-  /// Find the layer's state and it's index in the list from it's id
-  const findLayer = useCallback(
-    (layerId: string): [number, Layer] => {
-      for (const [i, layer] of layers.entries()) {
-        if (layer.id === layerId) {
-          return [i, layer];
-        }
-      }
-
-      throw Error(`Could not find layer with id ${layerId}`);
-    },
-    [layers],
-  );
-
-  const updateLayer = useCallback(
-    (
-      layerId: LayerId,
-      callback: LayerUpdateCallback,
-      virtual: boolean = false,
-    ) => {
-      const [i, layer] = findLayer(layerId);
-      const newLayer = callback(layer);
-
-      const fun = virtual ? setVirtualLayers : setLayers;
-
-      fun((prev: Layer[]) => [
-        ...prev.slice(0, i),
-        newLayer,
-        ...prev.slice(i + 1),
-      ]);
-    },
-    [findLayer, setLayers, setVirtualLayers],
-  );
-
-  const editSelectedLayers = (
-    callback: LayerUpdateCallback,
-    virtual: boolean = false,
-  ) => {
-    const fun = virtual ? setVirtualLayers : setLayers;
-    fun((prev) => {
-      return prev.map((layer) => {
-        if (!layer.isSelected) {
-          return layer;
-        }
-        return callback(layer);
-      });
-    });
   };
 
   return (
